Move icons components to component/icons
