+++ src/components/icons/IconInfo.svelte
