+++ src/components/icons/IconNotification.svelte
