+++ src/components/icons/IconPencil.svelte
