+++ src/components/icons/IconProfileArrow.svelte
