+++ src/components/icons/IconSearch.svelte
