+++ src/components/icons/IconUser.svelte
